+++ cnccode/business_classes/BURenQuotation.inc.php
@@ -31,22 +31,35 @@ class BURenQuotation extends Business
     function updateRenQuotation(&$dsData)
     {
         $this->setMethodName('updateRenQuotation');
-        $this->updateDataAccessObject($dsData, $this->dbeRenQuotation);
+        $this->updateDataAccessObject(
+            $dsData,
+            $this->dbeRenQuotation
+        );
 
         return TRUE;
     }
 
-    function getRenQuotationByID($ID, &$dsResults)
+    function getRenQuotationByID($ID,
+                                 &$dsResults
+    )
     {
         $this->dbeJRenQuotation->setPKValue($ID);
         $this->dbeJRenQuotation->getRow();
-        return ($this->getData($this->dbeJRenQuotation, $dsResults));
+        return ($this->getData(
+            $this->dbeJRenQuotation,
+            $dsResults
+        ));
     }
 
-    function getAll(&$dsResults, $orderBy = false)
+    function getAll(&$dsResults,
+                    $orderBy = false
+    )
     {
         $this->dbeJRenQuotation->getRows($orderBy);
-        return ($this->getData($this->dbeJRenQuotation, $dsResults));
+        return ($this->getData(
+            $this->dbeJRenQuotation,
+            $dsResults
+        ));
     }
 
     function deleteRenQuotation($ID)
@@ -78,10 +91,22 @@ class BURenQuotation extends Business
 
         $dsCustomerItem->setUpdateModeInsert();
 
-        $dsCustomerItem->setValue('customerItemID', 0);
-        $dsCustomerItem->setValue('customerID', $customerID);
-        $dsCustomerItem->setValue('itemID', $itemID);
-        $dsCustomerItem->setValue('siteNo', $siteNo);
+        $dsCustomerItem->setValue(
+            'customerItemID',
+            0
+        );
+        $dsCustomerItem->setValue(
+            'customerID',
+            $customerID
+        );
+        $dsCustomerItem->setValue(
+            'itemID',
+            $itemID
+        );
+        $dsCustomerItem->setValue(
+            'siteNo',
+            $siteNo
+        );
 
         $dsCustomerItem->post();
 
@@ -92,13 +117,34 @@ class BURenQuotation extends Business
 
         $this->dbeRenQuotation->getRow($customerItemID);
 
-        $this->dbeRenQuotation->setValue('customerItemID', $customerItemID);
-        $this->dbeRenQuotation->setValue('renQuotationTypeID', $renQuotationTypeID);
-        $this->dbeRenQuotation->setValue('startDate', date('Y-m-d'));
-        $this->dbeRenQuotation->setValue('qty', $qty);
-        $this->dbeRenQuotation->setValue('salePrice', $salePrice);
-        $this->dbeRenQuotation->setValue('costPrice', $costPrice);
-        $this->dbeRenQuotation->setValue('grantNumber', '');
+        $this->dbeRenQuotation->setValue(
+            'customerItemID',
+            $customerItemID
+        );
+        $this->dbeRenQuotation->setValue(
+            'renQuotationTypeID',
+            $renQuotationTypeID
+        );
+        $this->dbeRenQuotation->setValue(
+            'startDate',
+            date('Y-m-d')
+        );
+        $this->dbeRenQuotation->setValue(
+            'qty',
+            $qty
+        );
+        $this->dbeRenQuotation->setValue(
+            'salePrice',
+            $salePrice
+        );
+        $this->dbeRenQuotation->setValue(
+            'costPrice',
+            $costPrice
+        );
+        $this->dbeRenQuotation->setValue(
+            'grantNumber',
+            ''
+        );
 
         $this->dbeRenQuotation->updateRow();
 
@@ -114,10 +160,10 @@ class BURenQuotation extends Business
 
         $hdrs =
             array(
-                'From' => $senderEmail,
-                'To' => $toEmail,
-                'Subject' => 'Quotation Renewals Due Today',
-                'Date' => date("r"),
+                'From'         => $senderEmail,
+                'To'           => $toEmail,
+                'Subject'      => 'Quotation Renewals Due Today',
+                'Date'         => date("r"),
                 'Content-Type' => 'text/html; charset=UTF-8'
             );
 
@@ -147,9 +193,9 @@ class BURenQuotation extends Business
 
         $mime_params = array(
             'text_encoding' => '7bit',
-            'text_charset' => 'UTF-8',
-            'html_charset' => 'UTF-8',
-            'head_charset' => 'UTF-8'
+            'text_charset'  => 'UTF-8',
+            'html_charset'  => 'UTF-8',
+            'head_charset'  => 'UTF-8'
         );
         $body = $buMail->mime->get($mime_params);
 
@@ -173,10 +219,10 @@ class BURenQuotation extends Business
 
         $hdrs =
             array(
-                'From' => $senderEmail,
-                'To' => $toEmail,
-                'Subject' => 'Quotation Renewals Generated in Past 2 Weeks',
-                'Date' => date("r"),
+                'From'         => $senderEmail,
+                'To'           => $toEmail,
+                'Subject'      => 'Quotation Renewals Generated in Past 2 Weeks',
+                'Date'         => date("r"),
                 'Content-Type' => 'text/html; charset=UTF-8'
             );
 
@@ -206,9 +252,9 @@ class BURenQuotation extends Business
 
         $mime_params = array(
             'text_encoding' => '7bit',
-            'text_charset' => 'UTF-8',
-            'html_charset' => 'UTF-8',
-            'head_charset' => 'UTF-8'
+            'text_charset'  => 'UTF-8',
+            'html_charset'  => 'UTF-8',
+            'head_charset'  => 'UTF-8'
         );
 
         $body = $buMail->mime->get($mime_params);
@@ -253,7 +299,9 @@ class BURenQuotation extends Business
         $dbeCustomer = new DBECustomer ($this);
 
         $previousCustomerID = 99999;
+        $previousRenQuotationType = null;
 
+        $custItemsSharingSalesOrder = [];
         while ($this->dbeJRenQuotation->fetchNext()) {
             ?>
             quotation
@@ -265,26 +313,64 @@ class BURenQuotation extends Business
                 /*
                  * Group many renewals for same customer under one quote
                  */
-                if ($previousCustomerID != $dbeJCustomerItem->getValue('customerID')) {
+                if ($previousCustomerID != $dbeJCustomerItem->getValue(
+                        'customerID'
+                    ) || $previousRenQuotationType != $dbeJCustomerItem->getValue(
+                        DBEJCustomerItem::renQuotationTypeID
+                    )) {
                     /*
                      *  create order header
                      */
                     $dbeCustomer->getRow($dbeJCustomerItem->getValue('customerID'));
-                    $this->getData($dbeCustomer, $dsCustomer);
+                    $this->getData(
+                        $dbeCustomer,
+                        $dsCustomer
+                    );
 
-                    $buSalesOrder->initialiseQuote($dsOrdhead, $dsOrdline, $dsCustomer);
+                    $buSalesOrder->initialiseQuote(
+                        $dsOrdhead,
+                        $dsOrdline,
+                        $dsCustomer
+                    );
 
                     $line = -1;    // initialise sales order line seq
 
-                    $quotationIntroduction = 'Please find detailed below a quote for your ' . $this->dbeJRenQuotation->getValue('type') . ' renewal.';
+                    $quotationIntroduction = 'Please find detailed below a quote for your ' . $this->dbeJRenQuotation->getValue(
+                            'type'
+                        ) . ' renewal.';
 
                     $dbeOrdhead->getRow($dsOrdhead->getValue('ordheadID'));
 
-                    $dbeOrdhead->setValue('quotationSubject', $quotationSubject);
-                    $dbeOrdhead->setValue('quotationIntroduction', $quotationIntroduction);
+                    $dbeOrdhead->setValue(
+                        'quotationIntroduction',
+                        $quotationIntroduction
+                    );
+
+                    $dbeOrdhead->setValue(
+                        DBEOrdhead::quotationSubject,
+                        ucwords(
+                            $this->dbeJRenQuotation->getValue(
+                                'type'
+                            ) . ' renewal.'
+                        )
+                    );
 
                     $dbeOrdhead->updateRow();
+                    foreach ($custItemsSharingSalesOrder as $custItemID) {
+                        $dbeRenQuotationUpdate->setValue(
+                            'customerItemID',
+                            $custItemID
+                        );
+                        $dbeRenQuotationUpdate->getRow();
+                        $dbeRenQuotationUpdate->setValue(
+                            DBECustomerItem::ordheadID,
+                            $dbeOrdhead->getValue(DBEOrdhead::ordheadID)
+                        );
+                    }
+
+                    $custItemsSharingSalesOrder = [];
                 }
+                $custItemsSharingSalesOrder[] = $this->dbeJRenQuotation->getValue(DBEJRenQuotation::customerItemID);
 
                 $line++;
 
@@ -294,23 +380,64 @@ class BURenQuotation extends Business
                     $comment = $this->dbeJRenQuotation->getValue('comment');
                 } else {
                     $comment = $this->dbeJRenQuotation->getValue('type') . ' Renewal';
-//					$comment = $quotationSubject;
                 }
 
-                $dbeOrdline->setValue('lineType', 'C');
-                $dbeOrdline->setValue('renewalCustomerItemID', '');
-                $dbeOrdline->setValue('ordheadID', $dsOrdhead->getValue('ordheadID'));
-                $dbeOrdline->setValue('customerID', $dsOrdhead->getValue('customerID'));
-                $dbeOrdline->setValue('itemID', 0);
-                $dbeOrdline->setValue('description', $comment);
-                $dbeOrdline->setValue('supplierID', '');
-                $dbeOrdline->setValue('sequenceNo', $line);
-                $dbeOrdline->setValue('lineType', 'C');
-                $dbeOrdline->setValue('qtyOrdered', 0); // default 1
-                $dbeOrdline->setValue('qtyDespatched', 0);
-                $dbeOrdline->setValue('qtyLastDespatched', 0);
-                $dbeOrdline->setValue('curUnitSale', 0);
-                $dbeOrdline->setValue('curUnitCost', 0);
+                $dbeOrdline->setValue(
+                    'lineType',
+                    'C'
+                );
+                $dbeOrdline->setValue(
+                    'renewalCustomerItemID',
+                    ''
+                );
+                $dbeOrdline->setValue(
+                    'ordheadID',
+                    $dsOrdhead->getValue('ordheadID')
+                );
+                $dbeOrdline->setValue(
+                    'customerID',
+                    $dsOrdhead->getValue('customerID')
+                );
+                $dbeOrdline->setValue(
+                    'itemID',
+                    0
+                );
+                $dbeOrdline->setValue(
+                    'description',
+                    $comment
+                );
+                $dbeOrdline->setValue(
+                    'supplierID',
+                    ''
+                );
+                $dbeOrdline->setValue(
+                    'sequenceNo',
+                    $line
+                );
+                $dbeOrdline->setValue(
+                    'lineType',
+                    'C'
+                );
+                $dbeOrdline->setValue(
+                    'qtyOrdered',
+                    0
+                ); // default 1
+                $dbeOrdline->setValue(
+                    'qtyDespatched',
+                    0
+                );
+                $dbeOrdline->setValue(
+                    'qtyLastDespatched',
+                    0
+                );
+                $dbeOrdline->setValue(
+                    'curUnitSale',
+                    0
+                );
+                $dbeOrdline->setValue(
+                    'curUnitCost',
+                    0
+                );
 
                 $dbeOrdline->insertRow();
 
@@ -320,21 +447,63 @@ class BURenQuotation extends Business
                  * Get stock category from item table
                  */
                 $buItem = new BUItem($this);
-                $buItem->getItemByID($dbeJCustomerItem->getValue('itemID'), $dsItem);
-                $dbeOrdline->setValue('stockcat', $dsItem->getValue('stockcat'));
+                $buItem->getItemByID(
+                    $dbeJCustomerItem->getValue('itemID'),
+                    $dsItem
+                );
+                $dbeOrdline->setValue(
+                    'stockcat',
+                    $dsItem->getValue('stockcat')
+                );
 
-                $dbeOrdline->setValue('renewalCustomerItemID', $this->dbeJRenQuotation->getValue('customerItemID'));
-                $dbeOrdline->setValue('ordheadID', $dsOrdhead->getValue('ordheadID'));
-                $dbeOrdline->setValue('customerID', $dsOrdhead->getValue('customerID'));
-                $dbeOrdline->setValue('itemID', $dbeJCustomerItem->getValue('itemID'));
-                $dbeOrdline->setValue('description', $dbeJCustomerItem->getValue('itemDescription'));
-                $dbeOrdline->setValue('supplierID', CONFIG_SALES_STOCK_SUPPLIERID);
-                $dbeOrdline->setValue('sequenceNo', $line);
-                $dbeOrdline->setValue('lineType', 'I');
-                $dbeOrdline->setValue('qtyOrdered', $this->dbeJRenQuotation->getValue('qty'));
-                $dbeOrdline->setValue('qtyDespatched', 0);
-                $dbeOrdline->setValue('curUnitSale', $this->dbeJRenQuotation->getValue('salePrice'));
-                $dbeOrdline->setValue('curUnitCost', $this->dbeJRenQuotation->getValue('costPrice'));
+                $dbeOrdline->setValue(
+                    'renewalCustomerItemID',
+                    $this->dbeJRenQuotation->getValue('customerItemID')
+                );
+                $dbeOrdline->setValue(
+                    'ordheadID',
+                    $dsOrdhead->getValue('ordheadID')
+                );
+                $dbeOrdline->setValue(
+                    'customerID',
+                    $dsOrdhead->getValue('customerID')
+                );
+                $dbeOrdline->setValue(
+                    'itemID',
+                    $dbeJCustomerItem->getValue('itemID')
+                );
+                $dbeOrdline->setValue(
+                    'description',
+                    $dbeJCustomerItem->getValue('itemDescription')
+                );
+                $dbeOrdline->setValue(
+                    'supplierID',
+                    CONFIG_SALES_STOCK_SUPPLIERID
+                );
+                $dbeOrdline->setValue(
+                    'sequenceNo',
+                    $line
+                );
+                $dbeOrdline->setValue(
+                    'lineType',
+                    'I'
+                );
+                $dbeOrdline->setValue(
+                    'qtyOrdered',
+                    $this->dbeJRenQuotation->getValue('qty')
+                );
+                $dbeOrdline->setValue(
+                    'qtyDespatched',
+                    0
+                );
+                $dbeOrdline->setValue(
+                    'curUnitSale',
+                    $this->dbeJRenQuotation->getValue('salePrice')
+                );
+                $dbeOrdline->setValue(
+                    'curUnitCost',
+                    $this->dbeJRenQuotation->getValue('costPrice')
+                );
 
                 $dbeOrdline->insertRow();
                 /**
@@ -347,22 +516,67 @@ class BURenQuotation extends Business
                      * Get stock category from item table
                      */
                     $buItem = new BUItem($this);
-                    $buItem->getItemByID(CONFIG_INSTALLATION_ITEMID, $dsItem);
-                    $dbeOrdline->setValue('stockcat', $dsItem->getValue('stockcat'));
+                    $buItem->getItemByID(
+                        CONFIG_INSTALLATION_ITEMID,
+                        $dsItem
+                    );
+                    $dbeOrdline->setValue(
+                        'stockcat',
+                        $dsItem->getValue('stockcat')
+                    );
 
-                    $dbeOrdline->setValue('renewalCustomerItemID', 0);
-                    $dbeOrdline->setValue('ordheadID', $dsOrdhead->getValue('ordheadID'));
-                    $dbeOrdline->setValue('customerID', $dsOrdhead->getValue('customerID'));
-                    $dbeOrdline->setValue('itemID', CONFIG_INSTALLATION_ITEMID);
-                    $dbeOrdline->setValue('description', 'Installation/Configuration');
-                    $dbeOrdline->setValue('supplierID', CONFIG_SALES_STOCK_SUPPLIERID);
-                    $dbeOrdline->setValue('sequenceNo', $line);
-                    $dbeOrdline->setValue('lineType', 'I');
-                    $dbeOrdline->setValue('qtyOrdered', 1);
-                    $dbeOrdline->setValue('qtyDespatched', 0);
-                    $dbeOrdline->setValue('qtyLastDespatched', 0);
-                    $dbeOrdline->setValue('curUnitCost', $dsItem->getValue('curUnitCost'));
-                    $dbeOrdline->setValue('curUnitCost', $dsItem->getValue('curUnitSale'));
+                    $dbeOrdline->setValue(
+                        'renewalCustomerItemID',
+                        0
+                    );
+                    $dbeOrdline->setValue(
+                        'ordheadID',
+                        $dsOrdhead->getValue('ordheadID')
+                    );
+                    $dbeOrdline->setValue(
+                        'customerID',
+                        $dsOrdhead->getValue('customerID')
+                    );
+                    $dbeOrdline->setValue(
+                        'itemID',
+                        CONFIG_INSTALLATION_ITEMID
+                    );
+                    $dbeOrdline->setValue(
+                        'description',
+                        'Installation/Configuration'
+                    );
+                    $dbeOrdline->setValue(
+                        'supplierID',
+                        CONFIG_SALES_STOCK_SUPPLIERID
+                    );
+                    $dbeOrdline->setValue(
+                        'sequenceNo',
+                        $line
+                    );
+                    $dbeOrdline->setValue(
+                        'lineType',
+                        'I'
+                    );
+                    $dbeOrdline->setValue(
+                        'qtyOrdered',
+                        1
+                    );
+                    $dbeOrdline->setValue(
+                        'qtyDespatched',
+                        0
+                    );
+                    $dbeOrdline->setValue(
+                        'qtyLastDespatched',
+                        0
+                    );
+                    $dbeOrdline->setValue(
+                        'curUnitCost',
+                        $dsItem->getValue('curUnitCost')
+                    );
+                    $dbeOrdline->setValue(
+                        'curUnitCost',
+                        $dsItem->getValue('curUnitSale')
+                    );
 
                     $dbeOrdline->insertRow();
                 }
@@ -373,35 +587,95 @@ class BURenQuotation extends Business
                 $description =
                     $this->dbeJRenQuotation->getValue('grantNumber') . ' ' .
                     ' Expires: ' . $this->dbeJRenQuotation->getValue('nextPeriodStartDate');
-                $dbeOrdline->setValue('lineType', 'C');
-                $dbeOrdline->setValue('renewalCustomerItemID', '');
-                $dbeOrdline->setValue('ordheadID', $dsOrdhead->getValue('ordheadID'));
-                $dbeOrdline->setValue('customerID', $dsOrdhead->getValue('customerID'));
-                $dbeOrdline->setValue('itemID', 0);
-                $dbeOrdline->setValue('description', $description);
-                $dbeOrdline->setValue('supplierID', '');
-                $dbeOrdline->setValue('sequenceNo', $line);
-                $dbeOrdline->setValue('lineType', 'C');
-                $dbeOrdline->setValue('qtyOrdered', 0); // default 1
-                $dbeOrdline->setValue('qtyDespatched', 0);
-                $dbeOrdline->setValue('qtyLastDespatched', 0);
-                $dbeOrdline->setValue('curUnitSale', 0);
-                $dbeOrdline->setValue('curUnitCost', 0);
+                $dbeOrdline->setValue(
+                    'lineType',
+                    'C'
+                );
+                $dbeOrdline->setValue(
+                    'renewalCustomerItemID',
+                    ''
+                );
+                $dbeOrdline->setValue(
+                    'ordheadID',
+                    $dsOrdhead->getValue('ordheadID')
+                );
+                $dbeOrdline->setValue(
+                    'customerID',
+                    $dsOrdhead->getValue('customerID')
+                );
+                $dbeOrdline->setValue(
+                    'itemID',
+                    0
+                );
+                $dbeOrdline->setValue(
+                    'description',
+                    $description
+                );
+                $dbeOrdline->setValue(
+                    'supplierID',
+                    ''
+                );
+                $dbeOrdline->setValue(
+                    'sequenceNo',
+                    $line
+                );
+                $dbeOrdline->setValue(
+                    'lineType',
+                    'C'
+                );
+                $dbeOrdline->setValue(
+                    'qtyOrdered',
+                    0
+                ); // default 1
+                $dbeOrdline->setValue(
+                    'qtyDespatched',
+                    0
+                );
+                $dbeOrdline->setValue(
+                    'qtyLastDespatched',
+                    0
+                );
+                $dbeOrdline->setValue(
+                    'curUnitSale',
+                    0
+                );
+                $dbeOrdline->setValue(
+                    'curUnitCost',
+                    0
+                );
 
                 $dbeOrdline->insertRow();
                 /*
                  * set generated date
                  */
-                $dbeRenQuotationUpdate->setValue('customerItemID', $this->dbeJRenQuotation->getPKValue());
+                $dbeRenQuotationUpdate->setValue(
+                    'customerItemID',
+                    $this->dbeJRenQuotation->getPKValue()
+                );
                 $dbeRenQuotationUpdate->getRow();
-                $dbeRenQuotationUpdate->setValue('dateGenerated', date(CONFIG_MYSQL_DATE));
+                $dbeRenQuotationUpdate->setValue(
+                    'dateGenerated',
+                    date(CONFIG_MYSQL_DATE)
+                );
                 $dbeRenQuotationUpdate->updateRow();
 
-                $previousCustomerID = $dbeJCustomerItem->getValue('customerID');
-
-//		Needs to be done at conversion to initial: $this->processQuotationRenewal( $dbeJCustomerItem->getValue ( 'customerItemID' ) );
+                $previousCustomerID = $dbeJCustomerItem->getValue(DBEJCustomerItem::customerID);
+                $previousRenQuotationType = $dbeJCustomerItem->getValue(DBEJCustomerItem::renQuotationTypeID);
             }
         }
+
+        foreach ($custItemsSharingSalesOrder as $custItemID) {
+            $dbeRenQuotationUpdate->setValue(
+                'customerItemID',
+                $custItemID
+            );
+            $dbeRenQuotationUpdate->getRow();
+            $dbeRenQuotationUpdate->setValue(
+                DBECustomerItem::ordheadID,
+                $dbeOrdhead->getValue(DBEOrdhead::ordheadID)
+            );
+        }
+
 
         if ($customerItemIDs) {
             return $dsOrdhead->getValue('ordheadID');
@@ -410,7 +684,10 @@ class BURenQuotation extends Business
 
     function processQuotationRenewal($customerItemID)
     {
-        $this->dbeRenQuotation->setValue('customerItemID', $customerItemID);
+        $this->dbeRenQuotation->setValue(
+            'customerItemID',
+            $customerItemID
+        );
         $this->dbeRenQuotation->getRowsByColumn('customerItemID');
 
         if ($this->dbeRenQuotation->fetchNext()) {